fix: address PR review — cross-product bug, IN clause, refactor
- Fix cross-product false-positives: use or(and(eq(siren), eq(year)))
  pair conditions instead of separate IN clauses for (siren, year) queries
- Fix unsafe IN clause: use inArray() from drizzle-orm for declarationIds
- Fix misleading 400 error: generic message covers both date_begin/date_end
- Extract business logic from route.ts into modules/export/fetchDeclarations.ts
- Rewrite tests to mock the module instead of DB internals (no more
  fragile selectCallCount global state)
- Add test for invalid date_end and date_end passthrough
